rules is 40% refactored
What reamins? all documentation needs to be read, see what the db doesnt store and if it could store it, if it can have to fix indexer and extrators, after that? update all files again... compare it to the new list of things that cant be stored as db, convert those to basic ast, potentially regex patterns to not lose coverage and go from there.... not fun...